refactor(search-bar): migrate styles to styled-components
