+++ project-movie-app/src/App.tsx
@@ -1,5 +1,5 @@
 import { Header } from './Header'
-import { SearchBar } from "./Content/SearchBar"
+import { SearchBar } from "./Content/SearchBar/SearchBar"
 import { CardContainer } from './Content/CardContainer'
 import { Button } from './Button'
 import { Footer } from "./Footer"
