+++ project-movie-app/src/Content/SearchBar/style.ts
@@ -1,24 +1,32 @@
-/* search bar */
+import styled from "styled-components"
 
-.search-bar {
+export const SearchBarWrapper = styled.div`
     display: flex;
     align-items: center;
     justify-content: center;
     gap: 15px;
     margin: 30px 0;
-}
-
-.search-bar label {
+`
+export const Label = styled.label`
     font-size: 16px;
     font-weight: 500;
     white-space: nowrap;
-}
+`
 
-.search-bar input[type="text"] {
+export const Input = styled.input`
     width: 100%;
     max-width: 400px;
     padding: 12px;
     font-size: 16px;
     border: none;
     border-radius: 4px;
-}
+
+    &:focus {
+        outline: none;
+    }
+
+    &:focus-visible {
+        outline: 2px solid white;
+        outline-offset: 4px;
+    }
+`
