+++ project-movie-app/src/Content/SearchBar/types.ts
